@@ -1,94 +1,98 @@
 {
-  "name": "angular-portfolio",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
-    "start": "nx serve",
-    "build": "nx build",
-    "test": "nx test",
-	 "web:compodoc:serve":" nx run mtg-app:compodoc --serve",
-	 "api:compodoc:serve":" nx run api:compodoc --serve",
-    "seed:config": "ts-node ./node_modules/typeorm-seeding/dist/cli.js config",
-    "seed:run": "ts-node --project apps/api/tsconfig.app.json ./node_modules/typeorm-seeding/dist/cli.js seed"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~13.1.0",
-    "@angular/cdk": "^13.1.1",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
-    "@angular/flex-layout": "^13.0.0-beta.36",
-    "@angular/forms": "~13.1.0",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
-    "@datorama/akita": "^7.0.1",
-    "@datorama/akita-ng-entity-service": "^7.0.0",
-    "@datorama/akita-ng-router-store": "^7.0.0",
-    "@datorama/akita-ngdevtools": "^7.0.0",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/config": "^1.1.5",
-    "@nestjs/core": "^8.0.0",
-    "@nestjs/jwt": "^8.0.0",
-    "@nestjs/passport": "^8.0.1",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nestjs/swagger": "^5.1.5",
-    "@nestjs/typeorm": "^8.0.2",
-    "@nrwl/angular": "13.4.1",
-    "bcrypt": "^5.0.1",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.13.2",
-    "dotenv": "^10.0.0",
-    "jwt-decode": "^3.1.2",
-    "passport": "^0.5.2",
-    "passport-jwt": "^4.0.0",
-    "pg": "^8.7.1",
-    "primeicons": "^5.0.0",
-    "primeng": "^13.0.3",
-    "reflect-metadata": "^0.1.13",
-    "rxjs": "~7.4.0",
-    "swagger-ui-express": "^4.3.0",
-    "tslib": "^2.0.0",
-    "typeorm": "^0.2.41",
-    "typeorm-seeding": "^1.6.1",
-    "zone.js": "~0.11.4"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.0",
-    "@angular-eslint/eslint-plugin": "~13.0.1",
-    "@angular-eslint/eslint-plugin-template": "~13.0.1",
-    "@angular-eslint/template-parser": "~13.0.1",
-    "@angular/cli": "~13.1.0",
-    "@angular/compiler-cli": "~13.1.0",
-    "@angular/language-service": "~13.1.0",
-    "@compodoc/compodoc": "^1.1.15",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "13.4.1",
-    "@nrwl/cypress": "13.4.1",
-    "@nrwl/eslint-plugin-nx": "13.4.1",
-    "@nrwl/jest": "13.4.1",
-    "@nrwl/linter": "13.4.1",
-    "@nrwl/nest": "13.4.1",
-    "@nrwl/node": "13.4.1",
-    "@nrwl/tao": "13.4.1",
-    "@nrwl/workspace": "13.4.1",
-    "@twittwer/compodoc": "^1.6.6",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "~5.3.0",
-    "@typescript-eslint/parser": "~5.3.0",
-    "cypress": "^9.1.0",
-    "eslint": "8.2.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "jest": "27.2.3",
-    "jest-preset-angular": "11.0.0",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.4.3"
-  }
+	"name": "angular-portfolio",
+	"version": "0.0.0",
+	"license": "MIT",
+	"scripts": {
+		"ng": "nx",
+		"postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
+		"start": "nx serve",
+		"build": "nx build",
+		"test": "nx test",
+		"web:compodoc:serve": " nx run mtg-app:compodoc --serve",
+		"api:compodoc:serve": " nx run api:compodoc --serve",
+		"seed:config": "ts-node ./node_modules/typeorm-seeding/dist/cli.js config",
+		"seed:run": "ts-node --project apps/api/tsconfig.app.json ./node_modules/typeorm-seeding/dist/cli.js seed"
+	},
+	"private": true,
+	"dependencies": {
+		"@angular/animations": "15.2.8",
+		"@angular/cdk": "15.2.8",
+		"@angular/common": "15.2.8",
+		"@angular/compiler": "15.2.8",
+		"@angular/core": "15.2.8",
+		"@angular/flex-layout": "15.0.0-beta.42",
+		"@angular/forms": "15.2.8",
+		"@angular/platform-browser": "15.2.8",
+		"@angular/platform-browser-dynamic": "15.2.8",
+		"@angular/router": "15.2.8",
+		"@datorama/akita": "8.0.0",
+		"@datorama/akita-ng-entity-service": "8.0.0",
+		"@datorama/akita-ng-router-store": "8.0.0",
+		"@datorama/akita-ngdevtools": "7.0.0",
+		"@nestjs/common": "9.4.0",
+		"@nestjs/config": "2.3.1",
+		"@nestjs/core": "9.4.0",
+		"@nestjs/jwt": "10.0.3",
+		"@nestjs/passport": "9.0.3",
+		"@nestjs/platform-express": "9.4.0",
+		"@nestjs/swagger": "6.3.0",
+		"@nestjs/typeorm": "9.0.1",
+		"@nrwl/angular": "15.9.2",
+		"bcrypt": "5.1.0",
+		"class-transformer": "^0.5.1",
+		"class-validator": "^0.14.0",
+		"dotenv": "16.0.0",
+		"jwt-decode": "^3.1.2",
+		"passport": "^0.6.0",
+		"passport-jwt": "^4.0.1",
+		"pg": "^8.7.1",
+		"primeicons": "6.0.1",
+		"primeng": "15.4.1",
+		"reflect-metadata": "^0.1.13",
+		"rxjs": "7.8.0",
+		"swagger-ui-express": "^4.3.0",
+		"tslib": "^2.4.1",
+		"typeorm": "^0.3.15",
+		"zone.js": "0.12.0"
+	},
+	"devDependencies": {
+		"@angular-devkit/build-angular": "15.2.6",
+		"@angular-devkit/core": "15.2.6",
+		"@angular-devkit/schematics": "15.2.6",
+		"@angular-eslint/eslint-plugin": "15.0.0",
+		"@angular-eslint/eslint-plugin-template": "15.0.0",
+		"@angular-eslint/template-parser": "15.0.0",
+		"@angular/cli": "~15.0.0",
+		"@angular/compiler-cli": "15.2.8",
+		"@angular/language-service": "15.2.8",
+		"@compodoc/compodoc": "^1.1.15",
+		"@nestjs/schematics": "9.1.0",
+		"@nestjs/testing": "9.4.0",
+		"@nrwl/cli": "15.9.2",
+		"@nrwl/cypress": "15.9.2",
+		"@nrwl/eslint-plugin-nx": "15.9.2",
+		"@nrwl/jest": "15.9.2",
+		"@nrwl/linter": "15.9.2",
+		"@nrwl/nest": "15.9.2",
+		"@nrwl/node": "15.9.2",
+		"@nrwl/workspace": "15.9.2",
+		"@schematics/angular": "15.2.6",
+		"@twittwer/compodoc": "^1.6.6",
+		"@types/jest": "29.4.4",
+		"@types/node": "18.7.1",
+		"@typescript-eslint/eslint-plugin": "5.36.1",
+		"@typescript-eslint/parser": "5.36.1",
+		"cypress": "^9.1.0",
+		"eslint": "8.15.0",
+		"eslint-config-prettier": "8.1.0",
+		"eslint-plugin-cypress": "^2.10.3",
+		"jest": "29.4.3",
+		"jest-environment-jsdom": "29.4.3",
+		"jest-preset-angular": "13.0.0",
+		"nx": "15.9.2",
+		"prettier": "2.6.2",
+		"ts-jest": "29.0.5",
+		"ts-node": "10.9.1",
+		"typescript": "4.9.5"
+	}
 }
