@@ -63,7 +63,7 @@ export class AuthController {
 
 	// @Get(':id')
 	// findOne(@Param('id') id: string) {
-	//   return this.authService.findOne(+id);
+	//   return this.authService.findOneBy(+id);
 	// }
 
 	// @Patch(':id')
