@@ -39,7 +39,7 @@ const routes: Routes = [
 ];
 
 @NgModule({
-	imports: [RouterModule.forRoot(routes, { enableTracing: false, relativeLinkResolution: 'legacy' })],
+	imports: [RouterModule.forRoot(routes, { enableTracing: false })],
 	exports: [RouterModule],
 })
 export class AppRoutingModule { }
